requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
gast==0.2.2
google-auth==1.16.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imagesize==0.7.1
joblib==0.14.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.2.1
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.0.25
opt-einsum==3.2.1
pandas==0.24.2
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==3.12
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.19.1
scipy==1.4.1
seaborn==0.8.1
six==1.12.0
SQLAlchemy==1.2.1
tensorboard==2.0.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tqdm==4.39.0
urllib3==1.25.9
Werkzeug==1.0.0
wrapt==1.12.1
